// src/painter/painter.dto.ts
export class PainterDto {
    painter_code: string;    
    painter_name: string;
    nid : number;
    contact : string;
    address: string;
    progoti_status_id:number;
    progoti_status:string;
    status: string;   
    company_type: number;
    region: number;
    area: number;
    territory: number;
    cluster: number;
  }

  export class UpdatePainterDto {
    painter_code: string;    
    painter_name: string;
    nid : number;
    contact : string;
    address: string;
    progoti_status_id:number;
    progoti_status:string;
    status: string;   
    company_type: number;
    region: number;
    area: number;
    territory: number;
    cluster: number;
  }